refactor: 5장에 들어가기 앞서 JDBC와 JdbcTemplate의 User 분리
