+++ src/main/java/springbook/jdbc/JdbcContext.java
@@ -5,7 +5,6 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import javax.sql.DataSource;
-import springbook.domain.User;
 
 public class JdbcContext {
