+++ src/main/java/springbook/jdbctemplate/UserDao.java
@@ -3,7 +3,6 @@ package springbook.jdbctemplate;
 import javax.sql.DataSource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
-import springbook.domain.User;
 
 public class UserDao {
